+++ frontend/src/features/Home/HomeAside.tsx
@@ -1,5 +1,5 @@
 import UserCard from '@/components/UserCard';
-import TopReviewCard from '../temp/ui/TopReviewCard';
+import TopReviewCard from '../GapReview/ui/TopReviewCard';
 import RecentCareCard from '../Care/ui/RecentCareCard';
 export default function HomeAside() {
   return (
