+++ frontend/src/pages/Test/Jinyoung.tsx
@@ -3,8 +3,8 @@ import { CertificateCard } from '@/features/Home/ui/card/CertificateCard';
 import SectionBlock from '@/features/Home/ui/SectionBlock';
 import { certificateList } from '@/features/Home/data/cerificateList';
 
-import { reviewData } from '@/features/temp/data/reviewList';
-import ReviewCard from '@/features/temp/ui/ReviewCard';
+import { reviewData } from '@/features/GapReview/data/reviewList';
+import ReviewCard from '@/features/GapReview/ui/ReviewCard';
 
 import { TimelineCard } from '@/features/Home/ui/card/TimelineCard';
 export default function Jinyoung() {
